fix bug, update function
In the end of 2018, I was assigned to this "lottery project" for the company annual meeting. After few step of debuging and updating, here is the thing I made. And thanks to the original author Beau-zihan, the activity is archieving a great success by using it.(the boss added 3 more times' bouns of it)